Add demand-driven firm upsize mechanism
Profitable firms with sector demand > 90% can expand workforce by 5%
per month (8% probability). Symmetric with existing voluntary downsize.

New Decision.Upsize case in Firm decision tree, wired in fallbackDecision
before downsize/digiInvest paths. Additional Kahan precision fix:
split incomeShare/consShare in deposit computation to avoid catastrophic
cancellation.

+++ src/main/scala/com/boombustgroup/amorfati/engine/steps/BankUpdateStep.scala
@@ -397,7 +397,8 @@ object BankUpdateStep:
     val newDep = PLN(
       terms(
         b.deposits.toDouble,
-        (hhFlows.incomeShare - hhFlows.consShare).toDouble,
+        hhFlows.incomeShare.toDouble,
+        -hhFlows.consShare.toDouble,
         (investNetDepositFlow * ws).toDouble,
         (jstDepositChange * ws).toDouble,
         (in.s7.netDomesticDividends * ws).toDouble,
